api/requirements.txt
# assumes these are installed *in addition to* anaconda 4.6.7 + fastai 1.0 on Python 3.7+
autopep8==1.4.3
channels==2.1.7
channels-redis==2.3.3
Django==2.1.7
django-cors-headers==2.4.1
django-extensions==2.1.6
django-imagekit==4.0.2
django-nested-admin==3.1.3
djangorestframework==3.9.2
httpie==1.0.2
ipdb==0.11
ipython==7.3.0
pep8==1.7.1
psycopg2-binary==2.7.7
